services/rentals_service/requirements.txt
Django==5.2.7
djangorestframework==3.15.2
psycopg2-binary==2.9.10
Pillow==10.4.0
PyJWT==2.9.0
gunicorn==23.0.0
